gcp/requirements.txt
tensorflow
google-cloud-storage
Pillow
fastapi
uvicorn
python-multipart